implement new api, merge binary, make private
